moved blog into apps
