+++ apps/blog/tasks.py
@@ -1,5 +1,5 @@
 from celery import shared_task
-from .models import Image
+from .models.models import Image
 
 @shared_task
 def handle_image_upload(image_id):
